+++ config/test.exs
@@ -1,5 +1,6 @@
 use Mix.Config
 
 config :stellar_sdk,
-  http_client: Stellar.Horizon.HackneyMock,
+  http_client_impl: Stellar.Horizon.Client.CannedClientImpl,
+  http_client: Stellar.Horizon.Client.CannedHTTPClient,
   network: :test