+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.1.1 (05.11.2021)
 * Update library namespace from StellarSDK to Stellar.
 * Upgrade stellar_base dependency.
+* Improve the Horizon layer and the HTTP client implementation.
 
 ## 0.1.0 (18.08.2021)
 * Horizon HTTP client